+++ uCNC/src/core/motion_control.c
@@ -175,7 +175,7 @@ uint8_t mc_line(float *target, motion_data_t *block_data)
     }
 
     // check travel limits (soft limits)
-    if (!io_check_boundaries(target))
+    if (!kinematics_check_boundaries(target))
     {
         if (cnc_get_exec_state(EXEC_JOG))
         {
@@ -596,9 +596,10 @@ uint8_t mc_home_axis(uint8_t axis, uint8_t axis_limit)
     g_settings.limits_invert_mask ^= axis_limit;
     // io_set_homing_limits_filter(LIMITS_DUAL_MASK);//if axis pin goes off triggers
     cnc_unlock(true);
-    mc_line(target, &block_data);
     // flags homing clear by the unlock
     cnc_set_exec_state(EXEC_HOMING);
+    mc_line(target, &block_data);
+    
     if (itp_sync() != STATUS_OK)
     {
         return STATUS_CRITICAL_FAIL;