+++ uCNC/src/hal/kinematics/kinematic_corexy.c
@@ -29,16 +29,28 @@ void kinematics_init(void)
 
 void kinematics_apply_inverse(float *axis, int32_t *steps)
 {
-    steps[0] = (int32_t)lroundf(g_settings.step_mm[0] * (axis[AXIS_X] + axis[AXIS_Y]));
-    steps[1] = (int32_t)lroundf(g_settings.step_mm[1] * (axis[AXIS_X] - axis[AXIS_Y]));
-    steps[2] = (int32_t)lroundf(g_settings.step_mm[2] * axis[AXIS_Z]);
+    steps[0] = (int32_t)lroundf(g_settings.step_per_mm[0] * (axis[AXIS_X] + axis[AXIS_Y]));
+    steps[1] = (int32_t)lroundf(g_settings.step_per_mm[1] * (axis[AXIS_X] - axis[AXIS_Y]));
+
+#if AXIS_COUNT > 2
+    for (uint8_t i = 2; i < AXIS_COUNT; i++)
+    {
+        steps[i] = (int32_t)lroundf(g_settings.step_per_mm[i] * axis[i]);
+    }
+#endif
 }
 
 void kinematics_apply_forward(int32_t *steps, float *axis)
 {
-    axis[AXIS_X] = (float)(step_mm_inv[0] * 0.5f * (float)(steps[0] + steps[1]));
-    axis[AXIS_Y] = (float)(step_mm_inv[1] * 0.5f * (float)(steps[0] - steps[1]));
-    axis[AXIS_Z] = (float)(step_mm_inv[2] * steps[2]);
+    axis[AXIS_X] = (float)(0.5f * (float)(steps[0] + steps[1]) / g_settings.step_per_mm[0]);
+    axis[AXIS_Y] = (float)(0.5f * (float)(steps[0] - steps[1]) / g_settings.step_per_mm[1]);
+
+#if AXIS_COUNT > 2
+    for (uint8_t i = 2; i < AXIS_COUNT; i++)
+    {
+        axis[i] = (((float)steps[i]) / g_settings.step_per_mm[i]);
+    }
+#endif
 }
 
 uint8_t kinematics_home(void)
@@ -89,10 +101,9 @@ uint8_t kinematics_home(void)
     }
 #endif
 
-    // unlocks the machine to go to offset
-    cnc_clear_exec_state(EXEC_HOMING);
     cnc_unlock(true);
-
+    // flags homing clear by the unlock
+    cnc_set_exec_state(EXEC_HOMING);
     float target[AXIS_COUNT];
     motion_data_t block_data = {0};
     mc_get_position(target);
@@ -110,11 +121,18 @@ uint8_t kinematics_home(void)
     mc_line(target, &block_data);
     itp_sync();
 
+    // unlocks the machine to go to offset
+    cnc_clear_exec_state(EXEC_HOMING);
+
+#ifdef SET_ORIGIN_AT_HOME_POS
+    memset(target, 0, sizeof(target));
+#else
     for (uint8_t i = AXIS_COUNT; i != 0;)
     {
         i--;
-        target[i] += (!(g_settings.homing_dir_invert_mask & (1 << i)) ? 0 : g_settings.max_distance[i]);
+        target[i] = (!(g_settings.homing_dir_invert_mask & (1 << i)) ? 0 : g_settings.max_distance[i]);
     }
+#endif
 
     // reset position
     itp_reset_rt_position(target);
@@ -134,6 +152,30 @@ void kinematics_apply_reverse_transform(float *axis)
     /*
     Define your custom transform inverse operation
     */
+}
+
+bool kinematics_check_boundaries(float *axis)
+{
+    if (!g_settings.soft_limits_enabled || cnc_get_exec_state(EXEC_HOMING))
+    {
+        return true;
+    }
+
+    for (uint8_t i = AXIS_COUNT; i != 0;)
+    {
+        i--;
+#ifdef SET_ORIGIN_AT_HOME_POS
+        float value = !(g_settings.homing_dir_invert_mask & (1 << i)) ? -axis[i] : axis[i];
+#else
+        float value = axis[i];
+#endif
+        if (value > g_settings.max_distance[i] || value < 0)
+        {
+            return false;
+        }
+    }
+
+    return true;
 }
 
 #endif